# tests/fixtures/openapi_baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Infinite Canvas",
    "description": "Migration API — Phase 0 scaffold",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat",
        "operationId": "chat_api_chat_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/agent": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat Agent",
        "operationId": "chat_agent_api_chat_agent_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat Stream",
        "operationId": "chat_stream_api_chat_stream_post",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIReference": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "default": ""
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": ""
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": ""
          },
          "mime": {
            "type": "string",
            "title": "Mime",
            "default": ""
          }
        },
        "type": "object",
        "title": "AIReference"
      },
...
      "ChatRequest": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id",
            "default": ""
          },
          "message": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Message"
          },
          "system_prompt": {
            "type": "string",
            "title": "System Prompt",
            "default": ""
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "image_model": {
            "type": "string",
            "title": "Image Model",
            "default": ""
          },
          "image_provider": {
            "type": "string",
            "title": "Image Provider",
            "default": ""
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "default": "chat"
          },
          "size": {
            "type": "string",
            "title": "Size",
            "default": "1024x1024"
          },
          "quality": {
            "type": "string",
            "title": "Quality",
            "default": "auto"
          },
          "reference_images": {
            "items": {
              "$ref": "#/components/schemas/AIReference"
            },
            "type": "array",
            "title": "Reference Images"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "comfly"
          },
          "ms_model": {
            "type": "string",
            "title": "Ms Model",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}